chore: Split files in directories
