+++ src/handler/eth.transfers.spec.ts
@@ -1,6 +1,6 @@
 import { Finding, HandleTransaction, TransactionEvent } from "forta-agent";
 import provideETHTransferHandler from "./eth.transfers";
-import { createAddress, generalTestFindingGenerator, TestTransactionEvent } from "./tests.utils";
+import { createAddress, generalTestFindingGenerator, TestTransactionEvent } from "../test/tests.utils";
 import { toWei } from "web3-utils";
 
 describe("ETH Transfer Agent Tests", () => {
