--- src/tests.ts
@@ -1,23 +1,0 @@
-import { TestTransactionEvent, TestBlockEvent, runBlock, createAddress, Agent, TraceProps } from "./tests.utils";
-import { CallParams, MockEthersProvider, MockEthersSigner } from "./mock.utils";
-
-export {
-  createAddress,
-  TestTransactionEvent,
-  TestBlockEvent,
-  runBlock,
-  Agent,
-  TraceProps,
-  CallParams,
-  MockEthersProvider,
-  MockEthersSigner,
-};
-
-export default {
-  createAddress,
-  TestTransactionEvent,
-  TestBlockEvent,
-  runBlock,
-  MockEthersProvider,
-  MockEthersSigner,
-};
